Split server-misc.test.ts into two focused files
server-misc.test.ts (1285 lines) exceeded the 1200-line limit.

- server-misc-routing.test.ts (641 lines): security headers, content-type
  validation, routing utils (redirect, redirectResponse, getPrivateKey),
  CSRF validation, tracking param stripping, router patterns, CDN errors
- server-misc-admin-handlers.test.ts (644 lines): admin action handler
  factories (createActionHandler, withEntityLoader, etc.) and confirmation
  handler utilities (verifyIdentifier, verifyOrRedirect, etc.)
